Part 2 of fixing the boot code: gcc 3.4 fixes.
The whole problem seems to be size. Which is odd, because it is said
that size doesn't matter. Anyway... Add -Os to strategic places in the
makefile to have the final loader be as mall as possible. This seems
to be enough to make it work. For now... I think something is more
fundamentally wrong; or something more fundamental is wrong. Potato,
potaato.


git-svn-id: svn://svn.freebsd.org/base/head/sys/boot/ficl@134148 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f