Make command line parameter definitions more readable
This patch redefines PARAM macro so that the code is no longer the
argument of the macro. The new macro (renamed to PARAM_HANDLER) only
defines the header of the parameter handling function and the code
follows the header in the same way as normal functions are defined.

Besides this is IMHO more readable, Doxygen also parses this correctly
and put the PARAM_HANDLER macros into the list of functions.