novaboot: Disambiguate grub titles for serial line control
When there are several grub menu entries, where the title of one entry is
a prefix of another title (e.g. diskbench and diskbench-ramdisk) and
when $grub_keys is configured to use search key to locate the menu entry
(which is the default settings), then this search can result in selection
of the wrong entry.

To avoid this unwandted behavior, we mark the end of the grub titles with
semicolon and use the semicolon also in the grub keystrokes.