wvtest macros reworked so that they can be used everywhere
Earlier, the macros can only be used in methods of class that inherited
from WvTest. It was not possible to use WV* macros in static methods etc.

Now, the WV* macros can be used everywhere, where wvtest.h is included.
This is very useful for debugging: Instead of adding zillions of
printfs(), it is sufficient to wrap interesting code by WV* macros and
it automatically prints meaningful messages.