wvtest: Consider test as finished when scripting engine finishes
Since the scripting engine can now wait for children, there is an easy
way to detect when the whole test was finished - when the scripting
engine announces that the script is done.

This method is only used when the scripting engine actually waits for a
child. Otherwise, only resetting the machine is considered as a sign for
finished test.