Make specific Message class for every bus
In order to automatically generate a graph of connections between devices
and buses, it is useful when every bus uses its own and unique class of
messages.

In this patch we introduce the new classes that are inherited from from
the previously used classes and have exactly the same content, only the
type name is different.