Do not clone imgs submodules by default
Previously, my intention was to always create the imgs symlink in build
directory and clone the submodule only when 'scons test' is invoked, but
because scons does not distinguish between files in build directory and
in repositories, this didn't work as I expected and the submodule was
cloned always.

Now neither symlink nor the submodule is created/cloned by default.
Only when scons test is run.