Avoid unnecessary parent protocol handshaking when the service is not yet ready
When parent's get_portal() returns ERETRY and we are allowed to block,
after waking up we should return ERETRY to the caller instead of ENONE.
Returning ERETRY ensures that the caller (i.e. call_server()) calls us
immediately again, while returning ENONE causes it to try to call the
service, which fails so open_session() to the parent is called again and
finally get_portal() is retried.