wvtest: Try to better handle timeout in wvtestrun
In the original version, the child run in a separate process group so it
was possible to kill all childs and stay running. We cannot use the
separate process group because qemu serial redirection to stdout does
not work. This means that killing the child's process group also kills
us.

This problem is not solved by this patch. We only flush the output with
the hope that we see more output before we kill ourselves.