disk: Fix uninitialized variable that caused BIOS to fail
vbios_disk.cc checks the value of MessageDisk::error after
bus_disk.send(...). The error code was previously set in s0 as shown
below:

  msg->error = _mb->bus_disk.send(msg2) ? MessageDisk::DISK_OK : MessageDisk::DISK_STATUS_DEVICE;

but with the new disk service I forgot to initialize it. Now, it is
initialized always to DISK_OK and the error status is returned as the
return value of the receive() function.