vancouver: Fix on-demand disk initialization
We need to use bus_diskcommit to find out the number of disk users.

Note: Disk is required by vbios_disk, which is enabled by default so the
disk service is almost always needed. Otherwise, we cannot boot VM from
disk.

If the disk service is not available we only issue a warning instead of
panic, because booting the VM by other means that from disk could still
work.