Attempt to improve signal handling in wvtestrun
The aim of this is to get more useful messages when the test is
terminated by a signal (either Ctrl-C or by timeout). Handling this
correctly is not easy. The original wvtest solution was to run the
tested program as a background process group, which means that it is
possible to reliably kill all children by kill -$pgid. Since qemu cannot
run on background (it needs to access the terminal), we cannot do this.

Upon signal reception we still kill the whole process group, which may
result also in killing our parents (e.g. scons) and when child blocks
ignores SIGTEM, scons finishes before the test finishes, which is not
desired. The current solution seems to work in most cases, but it is not
bulletproof.