novaboot: Chnage default server from os to erwin
This is to remove the problem with booting stale binaries.
If the binaries are uploaded to os, it make take some time for NFS to
make the new versions visible on erwin, which runs the tftpd. By uploading
directly to erwin, the new versions will be hopefully seen by tftpd
immediately.