Fix tests that failed tonight
The problem: When a thread exits with ud2a (see ASMFUNCS), the kernel
prints an error message. This happens at the same time while S0 prints
"sc: done" message. Therefore those two messages are intermixed on the
serial line and WvTest harness reports timeout because it did not see the
"sc: done" message. Since I enabled serial output in the kernel on my test
box yesterday, this become a problem.

The fix: WvTest::exit() blocks forever so that ud2a in exit() is never
reached in tests.