wvtest: Do not kill the child (qemu) if WVTEST_NOKILL is set
Normally, when the test finishes, wvnulrun kills it. Typically,
qemu or novaboot is killed this way. Sometimes it is not desired
to kill the child. For exaple, when we want to test a server.
Novaboot is used to boot the server and once it is up, we exit and
start communicating with the server.

This feature will be used to allow running libvirt tests in qemu.