novaboot: Add a way for generating files on the fly by executing shell commands
This will eliminate the need for storing big files needed for testing
provided that they can be easily regenerated.

For example we might have a single initramfs template image for many tests
and for each test we will append to this template another init scripts
that are needed for the particular test.