Add test for starting and stopping an application by using config service
Basically, it works, but there are some strange things happening
when I try to allocate a watchdog thread on another CPU.

Another weird thing was that if the started and stopped app is not allowed
to access log service, which should not matter because the app does not
attempt to do so, then s0 crashed. Unfortunately (or maybe fortunately :-),
I cannot reproduce it now.