Update the tests to exercise NoXlate parent protocol
To keep the code of echo services and clients simple to actualy serve
for educational purposes, we compile the same source multiple times
to produce binaries that use different type of parent/service protocol.

The only exception is the header for the client side of the echo protocol,
which is copied to a separate header, because we want to use both versions
at the same sime in parentperf test.