doc: Add new org-to-doxygen converter
This one is based on the new export facility of Org and it is much
easier to modify than the previous org-lparse based one.

The converter is far from complete, but it is more or less sufficient
for the documentation we currently have.

The main reasons why I use org-mode for writing the documentation are:
- it is easy to change the structure
- it is easy to enter tables
- figures can be generated on the fly (e.g. dot)