doc: Replace SCons Doxygen builder
The downloaded builder never worked correctly. Its goal was to find
dependencies and it did that by parsing Doxyfile and searching for the
source files on the disk. Although the file list looked reasonable, SCons
somehou ignored it and never rebuilt the documentation when some file
changed.

This is now replaced by a simple Command builder that gets executed every
time without checking for dependencies. This is not a problem, because
documentation build has to be invoked explicitly by running 'scons doc'.