novaboot: Attempt to handle relay race condition
Currently, I'm not sure, whether it is really a race condition or the
relay hangs completely.

Note: This is an old patch from disk-donor branch. Currently, I think
that the relay was dead, when I originally wrote this. But this should
not do any harm so I leave it here.