wvtest: Run tests on a pty, as another process group
This is needed if we want to be able kill all processes spawned by the
test at one. Note that the process can easily "escape" from the process
group, but our tests are nice and don't escape :)

Pty is needed for qemu with -serial option. It doesn't run without tty.